Add files via upload
GCaMP-recording (cycle frequency,08/07/2024):
This Spike2-script measures the cycle frequency (in Hz) between two “peaks” (calcium spikes). Analyzed GCaMP-recording are imported as a TXT-file into Spike2 as a waveform-channel. After running the script, select the button “Import” to extract all calcium-spikes as event. To do this, adjust the horizontal cursor “baseline” to the baseline of the recording and horizontal cursor “peak” to the minimum peak to be detected. The script extracts the events based on the time of a peak in the waveform and the level specified by the two horizontal cursor. After this process, the event-channel has to be saved. “Analysis” is used to calculate the cycle frequency for each event in a TXT-file.